Moving similar capabilites to abstractions and adding unit tests.
